requirements.txt
asgiref          
crispy-bootstrap4
Django
django-allauth
django-cleanup
django-crispy-forms
django-crontab
django-dbbackup
django-htmx
django-jazzmin
django-phonenumber-field
djangorestframework
mysqlclient
google-auth
pillow
pip
pytz
sqlparse
tzdataq